chore: merge auth endpoints
